Src/graylog-snapshot-build-downloader/src/requirements.txt
PyYAML==6.0
requests==2.28.0
tqdm==4.65.0
urllib3==1.26.16
